Refactor out Protocols
This commit moves all Protocols out to a Protocol folder to aid the
restructuring of the app.

Issue: #84
