+++ SteamScout/Models/Protocols/PropertyListReadable.swift
